+++ celescope/tools/tag/mapping_tag.py
@@ -5,9 +5,6 @@ from celescope.tools import utils
 from celescope.tools import parse_chemistry
 from celescope.tools.step import Step, s_common
 
-# n_mismatch = 1 if n_tag_barcode > N_TAG_BARCODE_THRESHOLD else 2
-N_TAG_BARCODE_THRESHOLD = 10000
-
 
 def get_opts_mapping_tag(parser, sub_program):
     parser.add_argument(
@@ -20,8 +17,10 @@ def get_opts_mapping_tag(parser, sub_program):
     )
     parser.add_argument(
         "--barcode_fasta",
-        help="""Required. Tag barcode fasta file. It will check the mismatches between tag barcode 
-sequence in R2 reads with all tag barcode sequence in barcode_fasta. 
+        help="""Required. Tag barcode fasta file. Comma-separated if there are multiple `C` segments in fq_pattern.
+The number of fasta files must equal the number of `C` segments.
+
+It will check the mismatches between tag barcode sequence in R2 reads with all tag barcode sequence in barcode_fasta. 
 It will assign read to the tag with mismatch < threshold. 
 If no such tag exists, the read is classified as invalid.
 
@@ -33,6 +32,17 @@ You can find the example barcode fasta file under `celescope/data/Clindex` or `c
         "--linker_fasta",
         help="""Optional. If provided, it will check the mismatches between linker sequence in R2 reads 
 with all linker sequence in linker_fasta. If no mismatch < len(linker) / 10 + 1, the read is classified as invalid.
+""",
+    )
+    parser.add_argument(
+        "--mismatch",
+        help="""Maximum number of mismatches allowed between tag barcode in R2 reads and barcode_fasta.
+Comma-separated if there are multiple `C` segments (one value per segment).
+If a single value is given, it applies to all `C` segments.
+If not specified, it will be determined automatically for each segment based on the number of tag barcodes:
+  >100000: 0
+  >10000:  1
+  <=10000: 2
 """,
     )
     if sub_program:
@@ -73,18 +83,39 @@ class Mapping_tag(Step):
         self.fq_pattern = args.fq_pattern
         self.linker_fasta = args.linker_fasta
         self.barcode_fasta = args.barcode_fasta
+        self.mismatch_arg = args.mismatch
 
-        # process
+        # process pattern
         self.pattern_dict = parse_chemistry.parse_pattern(self.fq_pattern)
+        self.c_slices = self.pattern_dict["C"]
+        self.n_c = len(self.c_slices)
 
-        self.barcode_dict, self.barcode_length = utils.read_fasta(
-            self.barcode_fasta, equal=True
-        )
-        len_C = sum(x.stop - x.start for x in self.pattern_dict["C"])
-        if len_C != self.barcode_length:
-            raise ValueError(f"""The length of tag barcode in fq_pattern({len_C}) != 
-                length of tag barcode in barcode_fasta({self.barcode_length})""")
+        # parse barcode_fasta
+        fasta_list = [f.strip() for f in self.barcode_fasta.split(",")]
+        if len(fasta_list) != self.n_c:
+            raise ValueError(
+                f"Number of barcode_fasta files ({len(fasta_list)}) "
+                f"must equal number of `C` segments in fq_pattern ({self.n_c})."
+            )
 
+        # read barcode fastas
+        self.barcode_dict_list = []
+        self.barcode_length_list = []
+        for i, fasta in enumerate(fasta_list):
+            bc_dict, bc_len = utils.read_fasta(fasta, equal=True)
+            expected_len = self.c_slices[i].stop - self.c_slices[i].start
+            if bc_len != expected_len:
+                raise ValueError(
+                    f"Length of tag barcode in fasta[{i}] ({bc_len}) != "
+                    f"length of C[{i}] in fq_pattern ({expected_len})."
+                )
+            self.barcode_dict_list.append(bc_dict)
+            self.barcode_length_list.append(bc_len)
+
+        # parse mismatch
+        self.n_mismatch_list = self._parse_mismatch()
+
+        # linker
         if self.linker_fasta and self.linker_fasta != "None":
             self.linker_dict, self.linker_length = utils.read_fasta(
                 self.linker_fasta, equal=True
@@ -96,8 +127,8 @@ class Mapping_tag(Step):
         else:
             self.linker_dict, self.linker_length = {}, 0
 
-        # mismatch
-        self.mismatch_dict = self.get_tag_barcode_mismatch_dict()
+        # mismatch dicts per segment
+        self.mismatch_dict_list = self._get_tag_barcode_mismatch_dicts()
 
         # variables
         self.total_reads = 0
@@ -114,32 +145,66 @@ class Mapping_tag(Step):
         self.read_count_file = f"{self.outdir}/{self.sample}_read_count.tsv"
         self.invalid_barcode_file = f"{self.outdir}/{self.sample}_invalid_barcode.tsv"
 
+    def _parse_mismatch(self):
+        if self.mismatch_arg is not None:
+            mismatch_strs = [m.strip() for m in self.mismatch_arg.split(",")]
+            mismatch_vals = [int(m) for m in mismatch_strs]
+            if len(mismatch_vals) == 1 and self.n_c > 1:
+                mismatch_vals = mismatch_vals * self.n_c
+            if len(mismatch_vals) != self.n_c:
+                raise ValueError(
+                    f"Number of mismatch values ({len(mismatch_vals)}) "
+                    f"must equal number of `C` segments ({self.n_c})."
+                )
+        else:
+            mismatch_vals = []
+            for bc_dict in self.barcode_dict_list:
+                n_barcodes = len(bc_dict)
+                if n_barcodes > 100000:
+                    mismatch_vals.append(0)
+                elif n_barcodes > 10000:
+                    mismatch_vals.append(1)
+                else:
+                    mismatch_vals.append(2)
+
+        for i in range(self.n_c):
+            if mismatch_vals[i] > self.barcode_length_list[i]:
+                mismatch_vals[i] = self.barcode_length_list[i]
+
+        return mismatch_vals
+
     @utils.add_log
-    def get_tag_barcode_mismatch_dict(self):
-        mismatch_dict = {}
-        n_mismatch = 1 if len(self.barcode_dict) > N_TAG_BARCODE_THRESHOLD else 2
-        for seq_id, seq in self.barcode_dict.items():
-            for mismatch_seq in parse_chemistry.create_mismatch_seqs(
-                seq, max_mismatch=n_mismatch
-            ):
-                mismatch_dict[mismatch_seq] = seq_id
+    def _get_tag_barcode_mismatch_dicts(self):
+        mismatch_dict_list = []
+        for i in range(self.n_c):
+            mismatch_dict = {}
+            n_mismatch = self.n_mismatch_list[i]
+            for seq_id, seq in self.barcode_dict_list[i].items():
+                for mismatch_seq in parse_chemistry.create_mismatch_seqs(
+                    seq, max_mismatch=n_mismatch
+                ):
+                    mismatch_dict[mismatch_seq] = seq_id
+            mismatch_dict_list.append(mismatch_dict)
+        return mismatch_dict_list
 
-        return mismatch_dict
-
-    def check_barcode_with_mismatch(self, barcode, seq_barcode, umi):
+    def check_barcode_with_mismatch(self, barcode, seq_barcode_list, umi):
         """
         Args:
             barcode: cell barcode
-            seq_barcode: tag barcode sequence
+            seq_barcode_list: list of tag barcode sequences, one per C segment
             umi: UMI sequence
         """
-        if seq_barcode in self.mismatch_dict:
-            seq_id = self.mismatch_dict[seq_barcode]
-            self.res_dic[barcode][seq_id][umi] += 1
-            self.reads_mapped += 1
-        else:
-            self.reads_unmapped_invalid_barcode += 1
-            self.invalid_barcode_dict[seq_barcode] += 1
+        matched_names = []
+        for i, seq_barcode in enumerate(seq_barcode_list):
+            if seq_barcode not in self.mismatch_dict_list[i]:
+                self.reads_unmapped_invalid_barcode += 1
+                self.invalid_barcode_dict[seq_barcode] += 1
+                return
+            matched_names.append(self.mismatch_dict_list[i][seq_barcode])
+
+        tag_name = "_".join(matched_names)
+        self.res_dic[barcode][tag_name][umi] += 1
+        self.reads_mapped += 1
 
     def process_read(self):
         with pysam.FastxFile(self.fq) as infile:
@@ -153,8 +218,7 @@ class Mapping_tag(Step):
                 if self.linker_length != 0:
                     seq_linker = "".join(seq[x] for x in self.pattern_dict["L"])
 
-                if self.barcode_dict:
-                    seq_barcode = "".join(seq[x] for x in self.pattern_dict["C"])
+                seq_barcode_list = [seq[s] for s in self.c_slices]
 
                 # check linker
                 if self.linker_length != 0:
@@ -172,8 +236,8 @@ class Mapping_tag(Step):
                     self.reads_unmapped_invalid_linker += 1
                     continue
 
-                # check barcode
-                self.check_barcode_with_mismatch(barcode, seq_barcode, umi)
+                # check barcode per segment
+                self.check_barcode_with_mismatch(barcode, seq_barcode_list, umi)
 
     def write_files(self):
         # write dic to pandas df